+++ NEWS.md
@@ -1,5 +1,6 @@
-# pipload (development version)
-
+# pipload 0.1.4
+* Improve messages when `root_dir` is not available
+* Add function `add add_gls_to_env()` to be used by other PIP packages
 * add old povcalnet paths to retrieve master file vintages.
 
 # pipload 0.1.3